refactor(arch): Rename utils.verify_token to auth
