+++ app/routers/applications.py
@@ -20,10 +20,10 @@ from app.core.settings import app_settings
 from app.schema import api as ApiSchema
 from app.schema.api import ChangeEmail
 
+from ..auth import OCP_only, get_current_user, get_user
 from ..core.user_dependencies import CognitoClient, get_cognito_client
 from ..db.session import get_db, transaction_session
 from ..schema import core
-from ..utils.verify_token import OCP_only, get_current_user, get_user
 
 from fastapi import Depends, Query, status  # isort:skip # noqa
 from fastapi import Form, UploadFile  # isort:skip # noqa
